+++ lib/constants.dart
@@ -1,12 +1,17 @@
 import 'dart:ui';
 
 const Color blau = Color.fromARGB(255, 3, 88, 174);
+const Color blauTransparent = Color.fromARGB(200, 3, 88, 174);
 const Color rot = Color.fromARGB(255, 255, 95, 36);
+const Color rotTransparent = Color.fromARGB(200, 255, 95, 36);
 const Color silber = Color.fromARGB(255, 213, 215, 217);
 const Color silberTransparent = Color.fromARGB(220, 213, 215, 217);
 const Color gruen = Color.fromARGB(255, 12, 195, 141);
+const Color gruenTransparent = Color.fromARGB(200, 12, 195, 141);
 const Color gelb = Color.fromARGB(255, 255, 198, 0);
+const Color gelbTransparent = Color.fromARGB(200, 255, 198, 0);
 const Color violett = Color.fromARGB(255, 170, 109, 211);
+const Color violettTransparent = Color.fromARGB(200, 170, 109, 211);
 
 const baseUrl = 'https://zkmf2024-server.herokuapp.com';
 const cloudFlareUrl = 'https://imagedelivery.net/b5PKQi-eHYmFXihh4Gjd5g/';